Move cpp function/ file to fem/
