+++ cpp/dolfinx/fem/dolfin_fem.h
@@ -5,7 +5,7 @@ namespace dolfinx
 /*! \namespace dolfinx::fem
     \brief Finite element method functionality
 
-    Classes and algorithms for finite element method operations, e.g. assembly.
+    Classes and algorithms for finite element method spaces and operations.
 */
 }
 
@@ -17,6 +17,8 @@ namespace dolfinx
 #include <dolfinx/fem/DofMap.h>
 #include <dolfinx/fem/FiniteElement.h>
 #include <dolfinx/fem/Form.h>
+#include <dolfinx/fem/Function.h>
+#include <dolfinx/fem/FunctionSpace.h>
 #include <dolfinx/fem/SparsityPatternBuilder.h>
 #include <dolfinx/fem/assembler.h>
 #include <dolfinx/fem/utils.h>
