--- cpp/dolfinx/function/dolfin_function.h
@@ -1,17 +1,0 @@
-#pragma once
-
-namespace dolfinx
-{
-/*! \namespace dolfinx::function
-    \brief Functions tools, including FEM functions and pointwise defined
-   functions
-
-    This namespace provides classes for representing finite element functions,
-   and coefficient functions that appear in forms.
-*/
-}
-
-// DOLFINX function interface
-
-#include <dolfinx/function/Function.h>
-#include <dolfinx/function/FunctionSpace.h>
